fixes #12109 - removing owner_id presence validation
in rails 4, owner is not populated when creating a join model
(user.roles << role created a UserRole on the background)
in rails 3, this is not a problem, in rails 4, the presence validation
fails because no owner_id exists nor does an owner exist. at that
moment, there’s no need to do the uniqueness validation, as it’s the
user’s first roles. thus, this fix.