Fixes #12104 - prevent overflow in host list for various locales
The width has to be manually defined for the clipping to work properly. I prefer using percentages
to pixels when relevant so that enlarging the main container will cause the table to maintain
the same ratio (only place I used pixels was for the checkbox that has a fixed size).

The host group name clipping, which is the only column here that isn't automatically clipped
because of the special way in which nested host groups need to be clipped.
The whole clipping problem will be much simpler once we enlarge the entire content with the
Patternfly issue as clipping should almost never occur in most places.

Here are some before & after screenshots, problems were also found in German (buttons too large)
& Japanese (line breaking headers):

**Before:**
![screenshot from 2015-10-15 16-44-49](https://cloud.githubusercontent.com/assets/1540531/10515536/a56945e0-735c-11e5-8b44-46b79efda98c.png)
![screenshot from 2015-10-15 16-45-24](https://cloud.githubusercontent.com/assets/1540531/10515550/aaf4dd08-735c-11e5-94d5-9880aac12e37.png)

**After:**
![screenshot from 2015-10-15 16-40-16](https://cloud.githubusercontent.com/assets/1540531/10515562/b75b53d8-735c-11e5-8d8a-f9098cb49154.png)
![screenshot from 2015-10-15 16-41-26](https://cloud.githubusercontent.com/assets/1540531/10515565/b8e84ba2-735c-11e5-98aa-d3af21997e66.png)