Implement autocomplete UI component in AngularJS
This is a proof of concept implementation of the autocomplete UI
component in AngularJS. This makes use of the base AngularJS library
as well as the Angular UI Bootstrap projec that provides implementations
of Bootstrap widgets in Angular. This implementation and testing setup
is largely drawn from the Bastion plugin.