Fixes #12418 - Correct dashboard behaviour for updated gridster
Updated version of gridster changed the widget class from gs_w to gs-w:
https://github.com/ducksboard/gridster.js/commit/c1668d9ab6261d3c5dc4215abdb73ca516492c19
This led to a bit of breakage in css and js which is fixed by this commit.
There was also a change in the way the total width of the gridster
element was calculated that required a change to the initialization.

(cherry picked from commit e75ad04b0e45f15954eb45b09af1f484027db402)