fixes #12342 - unsaved (new or cloned) hostgroups keep their puppetclasses
Unsaved hostgroups didn't respond correctly to individual_puppetclasses
because that method expects some objects to be saved in the DB. In case
of selecting a new environment, that lead to puppetclasses disappearing.
Solution is - when the object is not saved yet, fallback on the
puppetclasses method.

(cherry picked from commit 654d3beadd98e49adfa0f868266140322ed0c500)