Fixes #8799 - Show override button when removing overriden global param
Previously, when removing an override for an existing parameter the
override button was not displayed again until the page was reloaded.
Also saved call to `mark_params_override()` when overriding a single
parameter since we already know which row needs to be marked as
overriden, and used some css to simplify `mark_params_override()`.

Note this code will be rendered useless once global parameters are
turned into lookup keys as they will use the same inline-editing
available for lookup keys, however until that change is made this will
improve usability of global parmeter overrides.