fixes #7517 - remove overridden class parameters from envs on import
When a class parameter is overridden and the class import indicates it
should be removed, now delete the EnvironmentClass association.

This also changes behaviour when an overridden parameter is removed from
all environments as it and any LookupValues will now be deleted too.
Previously only non-overridden parameters would be deleted.  This
matches the behaviour of classes being removed from all environments.