Fixes #12931 - Nested taxonomies show parent parameters
When assigning parameters at the top of the tree (for example, in
<region>) we cannot see these in the children. We can add values to
the children with the same name and this overwrites as expected. The
values parent location are available as parameters in the YAML files. We
should be able to view the parents parameters along with options to
override these.

Version-Release number of selected component (if applicable):

  How reproducible: 100%

  Steps to Reproduce:
  1. manage locations
  2. create a new location
  3. select a parent for the new location
  4. review the location ( you see no parameters)

Actual results:

Child location is separate from the parent.

Expected results:

Child should inherit from the parent including parameters

Additional info: