Fixes #12720 - Add tabs for puppet on proxy show page
This patch adds two tabs to the proxy show page related to puppet
1. Puppet - this tab is for proxies managing a puppet master. This tab
displays information about the puppet master, such as number of classes
in each environment, latest reports and so on.
2. Puppet CA - This tab incorporates the two seperate pages that were
already present for managing the puppet ca's certificates and autosign
entry. It also incorporated the new autosign form in a modal view.