fixes #13039 - remove TemplateKind DB calls from UnattendedCtlr
UnattendedController created actions/methods during initialisation per
known TemplateKind, relying on the DB being populated.  This caused
ordering issues when a plugin loaded the controller in a test
environment as the kinds weren't yet in the DB.  The actions then were
not created.

The per-TemplateKind actions have been simplified to one fixed action
which is routed in the same way.