fixes #13112 - add options to HostStatus#relevant? to optimise reports
To match other methods on HostStatus::Status instances, relevant? has a
new options argument which is used from the global status building to
optimise config report loading.  This prevents an N+1 query on the
hosts index page.

Previous methods defined with zero arguments are still supported with
a deprecation warning.