Fixes #13164 - Add view_params permission
A new view_params permission was added for parameters inheriting from
Parameter object. The only exception is global parameters, which are
already handled by filter for CommonParameter resource.

This new permissions is also automatically added to viewer and site
manager roles, as well as any other roles with any of the existing
Parameters permissions.

With the patch it's now also possible to use granular filters for all
parameters that Foreman supports.

Contributions from:
* Ori Rabin (orrabin) <orabin@redhat.com>
* Marek Hulán (ares) <mhulan@redhat.com>