Fixes #13171: Prevent frozen array error when plugins have assets
When the application is configured to serve static assets, the
public/asset paths of the plugins need to be added to the middleware
stack. However, this needs to occur prior to the middleware stack
being built. This is attainable by moving it into the main application
configuration logic.

This also includes an update to the sprockets manifest name that
appears to have changed with sprockets version updates. This means
that plugin assets and their manifests were not being loaded in
production environments.