Refs #12720 - Allow puppet CA cache to work for https proxies
When dumping CA certificates to cache, the status object is also dumped.
Since the status object points to the API resource, Marshal will try
dumping the API resource as well. This will fail for https proxies as
the SSL certificates cannot be dumped. This prevents the dumping of the
api link which will be recreated once the object is loaded back from
cache.