fixes #13440 - rename Host#model_name to #hardware_model_name
Rails 4.2 defines a model_name instance method which is used in routing
and called from link_to.  Rename our model_name methods to
hardware_model_name to prevent this overlapping and causing link_to
failures.

Host#hardware_model_name is provided as an equivalent getter to replace
it, and the #model_name= getter/setters are provided but are deprecated.
The deprecation period has been brought forward one release from the
usual, so as not to block an upgrade to Rails 4.2.