Fixes #13748 - Add initial javascript testing
Since our code base contains a good amount of javascript (4633 LOC in 76
    files), we need some testing framework to ensure we do not get
regressions on this code. Integration tests are slow to write and to
execute, and hopefully unit testing our javascript code will make it
more readable too.

Jasmine is a good testing framework with a great plugin ecosystem and
widely used in other projects for this same use case (Rails + js).