Fixes #13293 - Make all ActiveRecord threads share the same connection
Using transactional fixtures means "1 test" - "1 transaction" - the
reason we moved to DatabaseCleaner is that the 'server thread'
wouldn't be able to access the data in the 'test thread' as it is in a
transaction.

So we turned transactional_fixtures off, so data is
available to the server thread, and decided to clean the db with
DatabaseCleaner. That was very slow, and bugs were hard to find when
some integration tests would fail every now and then.

What we're doing here is to force ActiveRecord to share the same
connection between the 'server thread' and the 'test thread'. It's fine
because both don't request data from the db at the same time. Since data
isn't saved to db but stored in a transaction, the speed boost is very
noticeable too.