fixes #13804 - set enforce_available_locales in app config
Rails 4.1.9 in rails/rails@4399a23 sets I18n.enforce_available_locales
to the value in the Rails app config, which now defaults to true,
overwriting the false value set in the initialiser.  Set it in the app
config to ensure it remains off (since fast_gettext always sets the
best valid locale).

The translation of known locale names which triggers the error on en_GB
and pt_BR is also fixed, as fast_gettext uses underscores and i18n uses
hyphens.