fixes #10662 - set default_locale to 'en'
When a default locale isn't set and FastGettext.set_locale() is called
with an unknown locale, it picks the first available.  This depends on
the readdir order of locale/, which depends on filesystem behaviour.

Now when generating the default language API docs, they are always in
English and when accessing the web UI with an unknown language
preference, the 'en' locale is provided.