fixes #13897 - explicitly set limit on string columns
Rails 4.2 removes the default limit of 255 characters on PostgreSQL and
SQLite DB adapters, causing migrations on empty DBs to create columns
with no limit.  The MySQL adapter does not do this, as VARCHAR is
limited to 255 characters.

To keep the database schema identical no matter which version or adapter
is in use, specify the existing column limits explicitly on all
add_column calls.  This doesn't intend to resize any columns, which
should be done with additional DB migrations.