Fixes #5859 - don't rely on a mac address being present when overriding the conflicts
We tried to initiate the `dhcp_record` for checking if conflicts were
there. The problem was the mac address was not available at that stage
when using the compute resources. It also turns out there is no need
to initiate the dhcp_record at this stage, as we do that again while
actually removing the conflicts (and if no conflicts are there,
nothing happens anyway).