Fixes #13968 - Show errors when all features in proxy are not recognized
When using an old version of Foreman, say '1.10' against a newer version
of the proxy, say one that contains the 'logs' feature, the /features
call will return an array ['logs'].
If you try to add a smart-proxy that only has this feature, the check we
make `reply.is_a?(Array) and reply.any?` is not enough. We should verify
we have the features in the db.

The problem is:

- Try to add new smart-proxy with features ['logs'] to old foreman (1.10)
  without this feature.
- Submit, it won't save. But it won't throw any error either.

We should display an error and say 'the features in this proxy are not
recognized by foreman'.