fixes #14133 - update to rubocop 0.39.0
Replaces hirb-unicode with the maintained hirb-unicode-steakknife fork,
due to conflicting requirements on unicode-display_width with the older
hirb-unicode and new rubocop.

Rake is unpinned due to the newer rubocop's compatibility with it, and
parser unpinned as the newer rubocop+parser combination works. Custom
rake test tasks have warnings disabled to match Rails behaviour and
prevent lots of output.