Webpack POC - attempt at using webpack for asset packaging
This is an initial POC to see how difficult it would be to use webpack
with npm to manage our assets instead of gemified assets.
To use:
```
npm install
npm install webpack -g
webpack
```
This POC replaces a couple of vendored and gemified assets with npm
modules, which are packed together into one bundle for us to use.
It also includes POC of ES2015 transpiling using babel.
If we decide to go further down this path, we could potentially use the
webpack dev server to allow automatic refresh of assets, and to create
different bundles depending e.g. on locale to reduce the size of the
compiled application.js.