fixes #14338 - move session store configuration earlier
When a plugin has an initializer relying on :build_middleware_stack, the
middleware stack is constructed with the default session store (cookies)
as the store hasn't been configured by config/initializers/ yet.  Moving
the config into the main app config ensures it's set before the
middleware stack is first loaded.