Fixes #14410 - Raise error on pending migrations on production
*Problem*

During foreman-installer --upgrade, maybe there's some error that will
make migrations not run, such as proxy unavailable, etc.. In that case,
if the user checks the application, everything will look fine. However,
when migrations are pending, permissions from plugins will not be
loaded into Foreman::AccessControl: app/services/foreman/plugin.rb#L217

This is a problem for plugins. I can set the 'view_activation_keys'
permissions because it's on the db, but if it's not loaded, it's
useless. This means the user won't be able to use these permissions and
won't be able to access any plugin resource.

*Solution*

Raise a Foreman Exception with the details on how to move forward to
continue.