Fixes #8245 - Updates OS on Hostgroup update
Prior to this commit
On the new host page if host groups are updated
the hostgroup changes dont get reflected.

For example
If hostgroup A had Redhat 7.1 as OS
and hostgroup B had Redhat 7.2

Now on the New Host page if HG A got applied and then HG B
The operating system would still show up as RedHat 7.1

This commit addresses that issue by applying host group changes
completely for New Hosts.