Fixes #14769 - avoid template comments because of deface
After many hours spent on debugging I realized that the issue is that
when action view compiles the template, some lines are concatenated
together just split by ";". This becomes problem when ERB tag contained
commentary symbol #. The reason is that when deface override is applied
it first converts the template to tree using Nokogiri (see Parser.convert).

This returns different results on different versions of underlying library.
I didn't try to compile different versions but with libxml2-2.7.6-20.el6.x86_64 and
libxslt-1.1.26-2.el6_3.1.x86_64 it does not wrap the silent erb tag with
"\n" so as a result, this tag ends up on the same line as previous one.
The same nokogiri but on newer lib versions
(libxml2-2.9.1-5.el7_1.2.x86_64, libxslt-1.1.28-5.el7.x86_64) wrapped
it.