Fixes #14843 - Improve rake config error handling
Currently rake config doesn't give you any hint when you provide an
invalid setting. Instead, it should tell you what was wrong with the
input, and try to guide you to a valid one.

This commit provides error handling for wrong JSON values, and prints
what the error was if the setting could not be saved.