Fixes #14924 - Improve performance of taxonomy edit page
Non standard ids were not cached in TaxHost.
This led to multiple repeat queries in the database each time any of the
`*_ids` function were being called.
Used templated were calculated by going over all hosts in the taxonomy
and trying to lookup the provisioning template for each one, we now look
for all templates that match any combination of os, hostgroup and
environment for the taxonomy's hosts.
Hosts in Taxhost did not include the :interface table, leading to n+1
queries for delegated attributes such as subnets and proxies.