Fixes #13741 - Prevent hidden parameters from erasing newline
Switching the input from textarea to password leads to all newline
characters being stripped from the value. This PR attempts to prevent
this by changing the font of the textarea to a font that displays a
bullet instead of every character, thereby preventing the need to
change the input into a password field.