Fixes #15011 - don't expect the facet tables in migration
The problem was the
db/migrate/20140910153654_move_host_nics_to_interfaces.rb tries to
read the attributes that get extended by the facets, but the facets
table might not be ready yet.

To avoid unnecessary checks in production where the host.attributes
is used pretty heavily, we define the method with checks in migration
only.