Fixes #15077 - Error HTML tags are shown
On some custom errors we show, the alert_header helper is escaping the
HTML tags we put on the alert title.

http://imgur.com/cm6l5ZW

Obviously these shouldn't be escaped. Fix is as simple as marking the
string that puts the tags as html_safe

(cherry picked from commit 0300fadf46da1526fe428a2cdb7183b24b13605b)