Fixes #15066 - operating systems and ovirt < 3.6
The operating systems endpoint was added in ovirt 3.6
https://bugzilla.redhat.com/show_bug.cgi?id=1050243. Older versions
were failing getting 404 when provisioning.

This patch adds a check for this case. We also cache the operating
system data in compute resource - it saves one request + avoids
overloading logs with 404 when the operating systems are not supported
on ovirt side.

Finally, we show the information about operating systems availability
in the compute resource show page.

If somebody cached the data with ovirt 3.5 and then updates the ovirt
instance, they need to save the compute resource again to refresh the
cache.