Fixes #15137 - Find correct jnotify container
jnotify creates message container on first use and caches it
internally. When using turbolinks, page body is swapped
but jnotify has no clue about the change and appends the
message to the old container which means no message is
displayed. Giva Labs (the author) provides
plugin zipped without any instructions for contributors
or reference to the repository, that is the reason I
chose to modify jnotify directly.