Fixes #12193 - Stop using counter_cache for host[group] counts
This is aimed to prevent deadlocks caused by Rails' implementation of
counter_cache that is hit when multiple hosts are created or updated at
the same time. An added benefit is that the counters will now display
the number of hosts scoped to the current taxonomy and authorized for
the user to see. It also removes a lot of code used to workaround
various counter_cache issues.
The downsides are that sorting and filtering based on host counts will
no longer work, and that page load times increase slightly for pages
containing host counters, such as enviroments index.