Fixes #15140 - Correctly handle failed setting change
The jEditable library contains a bug that causes any failed changes to
settings to break the layout of the setting input in following changes.
Since the library is not maintained any more, we had to change our copy
of the library to fix the bug. This is an intermidiate fix until we
replace jeditable with a better, maintained solution.