fixes #15355 - close interface modal at cr change
When creating a new host and selecting the compute resource, an XHR
request is initiated. When a user opens the interface modal before the
XHR request finishes, the open modal is destroyed at the time the
request finishes.
This does not close the modal correctly and the backdrop is not hidden.
The user is now locked in and can not click anything.
This commit closes the modal properly when the XHR finishes.

(cherry picked from commit a45a87fe36ceab3c643955867356dc0f97dbe193)