fixes #15525 - always update facts
Previously, last_compile was checked to see if facts were uploaded in
the last minute and if so, ignore processing them.  This code dates back
to the very first tagged version of Foreman in git - but these days we
have multiple providers so this check doesn't make sense.  It's quite
conceivable that Salt/Chef/Puppet/Subscription Manager could upload
facts within such a short window - in fact, it does happen during
provisioning.