Fixes #15655 - Improve ptable migration
Update host and hostgroups in migration without running validation. Some
plugins failed to run the migration because Host or Hostgroups added
validations for fields that did not exist at the time this migration was
created.
This change also improves performance of the migration by using the
`update_all` method instead of looping over all elements.