Fixes #15718 - Correctly display hash parameters
Lookup values should dispaly the value before type cast for hash type
keys. Rails 4.2.6 introduced a change to the form helpers that only used
the value before type cast for user-provided input, so we need to
specify the correct value for the field instead of relying on the rails
helper to do it.