Fixes #15675 - Improve report expiry performance
Slowness was being caused by an un-needed `.joins` method, which led
rails to creating suboptimal queries. Also removed default ordering for
a bit of extra performance.

I tested on a largeish DB with 3M reports and 6M logs, time was cut by a
factor of 13. See the PR for query plan comparison.