fixes #15787 - update shoulda-matchers to 3.x
Contains changes to models for the new behaviour of the uniqueness
validation test, which saves a model with one value and then attempts
to validate a second.

NestedAncestryCommon's validation now doesn't fail when the specified
ancestry/parent_id is non-existent (shoulda-matchers generates an
invalid value).

Validations of title on NestedAncestryCommon models have been removed
as it's derived automatically from the name field, and can't be tested
independently of the model name. This in any case was a duplicate
validation of the name which is scoped to ancestry.

Host::Base has #dup implemented so the copied model has a valid primary
interface for #name= to work, matching the behaviour of #initialize.