fixes #10982 - configure RestClient log to 'proxy' logger
Request/response information is sent to the proxy logger, showing the
URLs called, headers, request body etc.

The X-Request-ID header is now also set in the proxy requests to a
random ID stored in the session, to prevent the user's real session ID
being hijacked from the proxy request data. The ID should be correlated
against the proxy logger.