Fixes #16025 - Fix ERB & params to import subnets from proxy
Before this commit, the ERB would just throw a 500 because it tried to
show @subnet.name in a template for a @subnet that was still undefined.

Aside from that, the create_multiple method in the controller didn't
send the right call to filter the params, hence making it impossible to
create the imported subnets even with the ERB fix.