Fixes #16082 - FactValue helper shouldn't assume valid .host
As described on #16081 - the fact_values helper is expecting
that "fact_value.host" to return a meaningful value all the time.

However for certain hosts (DiscoveredHost now, but it could be others in
the future) /fact_values shows a 500 because "fact_value.host" doesn't
exist.

I think it'd be reasonable to expect fact_value.host to exist for Hosts
in core, but we should show N/A if that cannot be found. This way the
page will be more resilient to manual db manipulation, data corruption,
or plain bugs like the one mentioned above.